src/requirements.txt
Flask==2.1.0
Flask-WTF==1.0.1
Werkzeug==2.1.1
requests==2.28.2
gunicorn==20.1.0
pytest==7.2.2
pytest-flask==1.2.0
python-dotenv==1.0.0
